requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: Google Sheets support
gspread>=6.0.0
google-auth>=2.23.0
